Se modifico el tipo de datos en la edad a TINYINT(3)

+++ database/models/Profile.js
@@ -15,7 +15,7 @@ module.exports = (sequelize, dataTypes) => {
             type: dataTypes.DATE
         },
         age: {
-            type: dataTypes.INTEGER
+            type: dataTypes.TINYINT(3)
         },
         name: {
             type: dataTypes.STRING
